chore(docs): Improved empty state UI positioning calculations in the IDE example
# Project State

## Current Focus
Improved empty state UI positioning calculations in the IDE example

## Context
The IDE example's empty state UI was previously using explicit type conversions (as usize) for positioning calculations, which could lead to potential precision issues. This change removes unnecessary type conversions to simplify the code while maintaining the same visual positioning.

## Completed
- [x] Removed redundant `as usize` conversions from empty state positioning calculations
- [x] Simplified empty state message and hint text positioning logic

## In Progress
- [x] No active work in progress

## Blockers
- None

## Next Steps
1. Verify the visual positioning remains consistent with the previous implementation
2. Consider additional UI improvements for the empty state (e.g., more visual cues)
---
category: chore
scope: docs
---

+++ examples/ide.rs
@@ -626,12 +626,12 @@ impl Widget for IdeApp {
             } else {
                 // Empty state - no tabs open
                 let empty_msg = " 󰈙 No file open ";
-                let empty_y = (editor_y + editor_content_h / 2) as usize;
-                let empty_x = (editor_x + 1 + (editor_w.saturating_sub(2) - empty_msg.len() as u16) / 2) as usize;
+                let empty_y = editor_y + editor_content_h / 2;
+                let empty_x = editor_x + 1 + (editor_w.saturating_sub(2) - empty_msg.len() as u16) / 2;
                 draw_text(&mut plane, empty_x, empty_y, empty_msg, t.fg_muted, t.bg, false);
                 let hint_msg = "Press Ctrl+O to open a file or Ctrl+T for a new tab";
                 let hint_y = empty_y + 2;
-                let hint_x = (editor_x + 1 + (editor_w.saturating_sub(2) - hint_msg.len() as u16) / 2) as usize;
+                let hint_x = editor_x + 1 + (editor_w.saturating_sub(2) - hint_msg.len() as u16) / 2;
                 draw_text(&mut plane, hint_x, hint_y, hint_msg, t.fg_subtle, t.bg, false);
             }
         }
